+++ test/test_edge_index.py
@@ -11,6 +11,7 @@ from torch_geometric import EdgeIndex
 from torch_geometric.edge_index import (
     SUPPORTED_DTYPES,
     ReduceType,
+    SortReturnType,
     _scatter_spmm,
     _torch_sparse_spmm,
     _TorchSPMM,
@@ -18,7 +19,6 @@ from torch_geometric.edge_index import (
 )
 from torch_geometric.profile import benchmark
 from torch_geometric.testing import (
-    disableExtensions,
     onlyCUDA,
     onlyLinux,
     withCUDA,
@@ -46,10 +46,17 @@ def test_basic(dtype, device):
     adj = EdgeIndex([[0, 1, 1, 2], [1, 0, 2, 1]], **kwargs)
     adj.validate()
     assert isinstance(adj, EdgeIndex)
+
     if torch_geometric.typing.WITH_PT112:
-        assert str(adj).startswith('EdgeIndex([[0, 1, 1, 2],')
+        assert str(adj).startswith('EdgeIndex([[0, 1, 1, 2],\n'
+                                   '           [1, 0, 2, 1]], ')
     else:
-        assert str(adj).startswith('tensor([[0, 1, 1, 2],')
+        assert str(adj).startswith('tensor([[0, 1, 1, 2],\n'
+                                   '        [1, 0, 2, 1]], ')
+    assert str(adj).endswith('sparse_size=(3, 3), nnz=4)')
+    assert (f"device='{device}'" in str(adj)) == adj.is_cuda
+    assert (f'dtype={dtype}' in str(adj)) == (dtype != torch.long)
+
     assert adj.dtype == dtype
     assert adj.device == device
     assert adj.sparse_size() == (3, 3)
@@ -88,6 +95,31 @@ def test_identity(dtype, device, is_undirected):
     assert out.sparse_size() == adj.sparse_size()
     assert out.sort_order == adj.sort_order
     assert out.is_undirected == adj.is_undirected
+
+
+@withCUDA
+@pytest.mark.parametrize('dtype', DTYPES)
+def test_sparse_tensor(dtype, device):
+    kwargs = dict(dtype=dtype, device=device, is_undirected=True)
+    adj = EdgeIndex([[0, 1, 1, 2], [1, 0, 2, 1]], sort_order='row', **kwargs)
+
+    out = EdgeIndex(adj.to_sparse_coo())
+    assert out.equal(adj)
+    assert out.sort_order == 'row'
+    assert out.sparse_size() == (3, 3)
+    assert out._indptr is None
+
+    out = EdgeIndex(adj.to_sparse_csr())
+    assert out.equal(adj)
+    assert out.sort_order == 'row'
+    assert out.sparse_size() == (3, 3)
+    assert out._indptr.equal(tensor([0, 1, 3, 4], device=device))
+
+    out = EdgeIndex(adj.to_sparse_csc())
+    assert out.equal(adj.sort_by('col')[0])
+    assert out.sort_order == 'col'
+    assert out.sparse_size() == (3, 3)
+    assert out._indptr.equal(tensor([0, 1, 3, 4], device=device))
 
 
 def test_set_tuple_item():
@@ -311,7 +343,7 @@ def test_sort_by(dtype, device, is_undirected):
     kwargs = dict(dtype=dtype, device=device, is_undirected=is_undirected)
     adj = EdgeIndex([[0, 1, 1, 2], [1, 0, 2, 1]], sort_order='row', **kwargs)
     out = adj.sort_by('row')
-    assert isinstance(out, torch.return_types.sort)
+    assert isinstance(out, SortReturnType)
     assert isinstance(out.values, EdgeIndex)
     assert not isinstance(out.indices, EdgeIndex)
     assert out.values.equal(adj)
@@ -319,7 +351,7 @@ def test_sort_by(dtype, device, is_undirected):
 
     adj = EdgeIndex([[0, 1, 2, 1], [1, 0, 1, 2]], **kwargs)
     out = adj.sort_by('row')
-    assert isinstance(out, torch.return_types.sort)
+    assert isinstance(out, SortReturnType)
     assert isinstance(out.values, EdgeIndex)
     assert not isinstance(out.indices, EdgeIndex)
     assert out.values[0].equal(tensor([0, 1, 1, 2], device=device))
@@ -369,6 +401,11 @@ def test_cat(dtype, device, is_undirected):
     assert out.sparse_size() == (4, 4)
     assert not out.is_sorted
     assert out.is_undirected == is_undirected
+
+    assert out._cat_metadata.nnz == [4, 4]
+    assert out._cat_metadata.sparse_size == [(3, 3), (4, 4)]
+    assert out._cat_metadata.sort_order == [None, None]
+    assert out._cat_metadata.is_undirected == [is_undirected, is_undirected]
 
     out = torch.cat([adj1, adj2, adj3], dim=1)
     assert out.size() == (2, 12)
@@ -1042,7 +1079,6 @@ def test_torch_script():
 
 
 @onlyLinux
-@disableExtensions
 @withPackage('torch>=2.1.0')
 def test_compile():
     import torch._dynamo as dynamo