+++ torch_geometric/sampler/neighbor_sampler.py
@@ -222,9 +222,11 @@ class NeighborSampler(BaseSampler):
                 self.row, self.colptr, self.perm = graph_store.csc()
 
             else:
-                self.node_types = list(
-                    set(attr.group_name for attr in attrs
-                        if isinstance(attr.group_name, NodeType)))
+                node_types = [
+                    attr.group_name for attr in attrs
+                    if isinstance(attr.group_name, str)
+                ]
+                self.node_types = list(set(node_types))
                 self.num_nodes = {
                     node_type: remote_backend_utils.size(*data, node_type)
                     for node_type in self.node_types
@@ -232,18 +234,27 @@ class NeighborSampler(BaseSampler):
                 self.edge_weight: Optional[Dict[EdgeType, Tensor]] = None
 
                 self.node_time: Optional[Dict[NodeType, Tensor]] = None
-                self.edge_time: Optional[Dict[NodeType, Tensor]] = None
+                self.edge_time: Optional[Dict[EdgeType, Tensor]] = None
 
-                # TODO Add support for edge-level temporal sampling.
                 if time_attr is not None:
                     for attr in time_attrs:  # Reset index for full data.
                         attr.index = None
+
                     time_tensors = feature_store.multi_get_tensor(time_attrs)
-                    self.node_time = {
-                        time_attr.group_name: time_tensor
-                        for time_attr, time_tensor in zip(
-                            time_attrs, time_tensors)
+                    time = {
+                        attr.group_name: time_tensor
+                        for attr, time_tensor in zip(time_attrs, time_tensors)
                     }
+
+                    group_names = [attr.group_name for attr in time_attrs]
+                    if all([isinstance(g, str) for g in group_names]):
+                        self.node_time = time
+                    elif all([isinstance(g, tuple) for g in group_names]):
+                        self.edge_time = time
+                    else:
+                        raise ValueError(
+                            f"Found time attribute '{time_attr}' for both "
+                            f"node-level and edge-level types")
 
                 # Conversion to/from C++ string type (see above):
                 self.to_rel_type = {k: '__'.join(k) for k in self.edge_types}